% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_make_eco_mat.R
\name{make_eco_mat}
\alias{make_eco_mat}
\title{Calculation of ecological relationships}
\usage{
make_eco_mat(growth_file, alpha = 0.05)
}
\arguments{
\item{growth_file}{a list, as exported by the function metabolic_interactions_with_MicrobiomeGS2 and contains:

a. The result of the biomass optimization of each model alone

b. The result of the biomass optimization and the metabolic interactions for each pair of models}

\item{alpha}{the margin in the relative growth (co_growth - alone_growth)/alone_growth that would be counted as a considerable change, default is 0.05}
}
\value{
(a list object in the R environment)

a. matrix of all_bacteria X all_bacteria with each cell represents the ecological relation between the two bacteria

b. matrix of the change in growth for all models in columns in the presence of all other models in rows
}
\description{
Ecological relationship between pairs of co-grown bacteria
based on the list that is exported by the function metabolic_interactions_with_MicrobiomeGS2
}
\details{
A reference table for developers:\tabular{lccr}{
   Species 1 \tab Species 2 \tab Sum \tab Relation \cr
   0 \tab 0 \tab 0 \tab Neutralism \cr
   0 \tab 1 \tab 1 \tab Commensalism \cr
   0 \tab -1 \tab -1 \tab Amensalism \cr
   1 \tab 1 \tab 2 \tab Mutualism \cr
   1 \tab -1 \tab 0 \tab Parasitism \cr
   -1 \tab -1 \tab 2 \tab Competition \cr
}
}
