% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_plot_relations.R
\name{plot_relations}
\alias{plot_relations}
\title{Plotting ecological relations}
\usage{
plot_relations(eco_mat)
}
\arguments{
\item{eco_mat}{A list, as exported by the function make_eco_mat. The matrix of all_models X all_models with each cell represents the ecological relation between the two strains.}
}
\value{
A plot of the number of pairs in each type of ecological relations from the eco_mat
}
\description{
To plot the number of pairs in each type of ecological relations as exported by the function make_eco_mat
}
