% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_relation_ratios.R
\name{relation_ratios}
\alias{relation_ratios}
\title{Transformation for the ecological relationships into log10 of the pairwise ratios}
\usage{
relation_ratios(relations_table)
}
\arguments{
\item{relations_table}{the ecological relation frequencies per sample, exported by the function "relation_per_sample"}
}
\value{
A data frame of the log10 transformed ecological relations ratios
}
\description{
Transformation for the ecological relationships into log10 of the pairwise ratios
}
