% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_metabolic_interactions_with_MicrobiomeGS2.R
\name{metabolic_interactions_with_MicrobiomeGS2}
\alias{metabolic_interactions_with_MicrobiomeGS2}
\title{Metabolic interactions with MicrobiomeGS2}
\usage{
metabolic_interactions_with_MicrobiomeGS2(list_of_models, cores, save_pair)
}
\arguments{
\item{list_of_models}{a NAMED list of models in sybil format on the same diet}

\item{cores}{the number of available cores}

\item{save_pair}{set T to save each pair of models on the working directory}
}
\value{
(to be returned as an object in the R environment)

a. the result of the biomass optimisation of each model alone

b. the result of the biomass optimisation and the metabolic interactions for each pair of models

=======================================

(to be saved as a RDS file inside the working directory, if save_pair set to T)

For each pair, a list with:

a.the joined model

b.the common exchange reactions

c.the fluxes of the meta-model

d.the metabolic interactions among the sub-models
}
\description{
Metabolic interactions and growth capacity of metabolic models
based on MicrobiomeGS2 and sybil packages
}
